refactor: UserRestoreScheduler 서비스단에서 호출해오는 형태로 리팩터링
