+++ src/main/java/com/study/petory/domain/user/controller/TestSchedulerController.java
@@ -100,7 +100,7 @@ public class TestSchedulerController {
 		// 현재 날짜를 임의로 설정해서 테스트 (예: 복구 예정 날짜인 "2025-06-18T00:00")
 		LocalDateTime simulatedNow = LocalDateTime.parse(date);
 
-		userRestoreScheduler.testRestoreSuspendedUsers(simulatedNow);
+		userSchedulerService.testRestoreSuspendedUsers(simulatedNow);
 
 		return CommonResponse.of(SuccessCode.RESTORED);
 	}
